fix(oob): expose parseInvitation as a util
