+++ packages/core/src/modules/oob/OutOfBandModule.ts
@@ -1,7 +1,7 @@
 import type { AgentMessage } from '../../agent/AgentMessage'
 import type { AgentMessageReceivedEvent } from '../../agent/Events'
 import type { Logger } from '../../logger'
-import type { ConnectionRecord, Routing } from '../../modules/connections'
+import type { ConnectionRecord, Routing, ConnectionInvitationMessage } from '../../modules/connections'
 import type { PlaintextMessage } from '../../types'
 import type { Key } from '../dids'
 import type { HandshakeReusedEvent } from './domain/OutOfBandEvents'
@@ -18,12 +18,7 @@ import { MessageSender } from '../../agent/MessageSender'
 import { createOutboundMessage } from '../../agent/helpers'
 import { ServiceDecorator } from '../../decorators/service/ServiceDecorator'
 import { AriesFrameworkError } from '../../error'
-import {
-  DidExchangeState,
-  HandshakeProtocol,
-  ConnectionInvitationMessage,
-  ConnectionsModule,
-} from '../../modules/connections'
+import { DidExchangeState, HandshakeProtocol, ConnectionsModule } from '../../modules/connections'
 import { DidCommMessageRepository, DidCommMessageRole } from '../../storage'
 import { JsonEncoder, JsonTransformer } from '../../utils'
 import { parseMessageType, supportsIncomingMessageType } from '../../utils/messageType'
@@ -280,18 +275,8 @@ export class OutOfBandModule {
    *
    * @returns OutOfBandInvitation
    */
-  public async parseInvitation(invitationUrl: string) {
-    const parsedUrl = parseUrl(invitationUrl).query
-    if (parsedUrl['oob']) {
-      const outOfBandInvitation = await OutOfBandInvitation.fromUrl(invitationUrl)
-      return outOfBandInvitation
-    } else if (parsedUrl['c_i'] || parsedUrl['d_m']) {
-      const invitation = await ConnectionInvitationMessage.fromUrl(invitationUrl)
-      return convertToNewInvitation(invitation)
-    }
-    throw new AriesFrameworkError(
-      'InvitationUrl is invalid. It needs to contain one, and only one, of the following parameters: `oob`, `c_i` or `d_m`.'
-    )
+  public async parseInvitation(invitationUrl: string): Promise<OutOfBandInvitation> {
+    return await this.parseInvitation(invitationUrl)
   }
 
   /**
